project/requirements.txt
flask
python-dotenv
mongoengine
pymongo
flask_wtf
wtforms
datetime
flask-bcrypt
flask-login
pillow
apscheduler
stripe